FELIX-6702
Jetty bundle
* org.apache.felix.jakarta.ee9.websocket.enable -> org.apache.felix.jakarta.websocket.enable
* org.apache.felix.jetty.ee9.websocket.enable -> org.apache.felix.jetty.websocket.enable

Jetty12 bundle
* org.apache.felix.jakarta.ee10.websocket.enable -> org.apache.felix.jakarta.websocket.enable
* org.apache.felix.jetty.ee10.websocket.enable -> org.apache.felix.jetty.websocket.enable

Classifiers:

* with-jetty-ee10-websockets -> with-jetty-websockets
* with-jakarta-ee10-websockets -> with-jakarta-websockets

Updated unit tests and README.md to reflect these changes.
