+++ http/jetty12/src/main/java/org/apache/felix/http/jetty/internal/JettyConfig.java
@@ -271,11 +271,11 @@ public final class JettyConfig
     /** Felix specific property to specify the default protocol when negotiation fails  */
     public static final String FELIX_JETTY_ALPN_DEFAULT_PROTOCOL = "org.apache.felix.jetty.alpn.defaultProtocol";
 
-    /** Felix specific property to control whether to enable the standard jakarta.websocket EE10 APIs provided by Jakarta WebSocket 2.1 */
-    public static final String FELIX_JAKARTA_EE10_WEBSOCKET_ENABLE = "org.apache.felix.jakarta.ee10.websocket.enable";
+    /** Felix specific property to control whether to enable the standard jakarta.websocket APIs provided by Jakarta WebSocket 2.1 */
+    public static final String FELIX_JAKARTA_WEBSOCKET_ENABLE = "org.apache.felix.jakarta.websocket.enable";
 
-    /** Felix specific property to control whether to enable they Jetty-specific EE10 WebSocket APIs */
-    public static final String FELIX_JETTY_EE10_WEBSOCKET_ENABLE = "org.apache.felix.jetty.ee10.websocket.enable";
+    /** Felix specific property to control whether to enable they Jetty-specific WebSocket APIs */
+    public static final String FELIX_JETTY_WEBSOCKET_ENABLE = "org.apache.felix.jetty.websocket.enable";
 
     private static String validateContextPath(String ctxPath)
     {
@@ -684,19 +684,19 @@ public final class JettyConfig
     }
 
     /**
-     * Returns <code>true</code> if jakarta EE10 websocket is configured to be used (
-     * {@link #FELIX_JAKARTA_EE10_WEBSOCKET_ENABLE})
+     * Returns <code>true</code> if jakarta websocket is configured to be used (
+     * {@link #FELIX_JAKARTA_WEBSOCKET_ENABLE})
      */
-    public boolean isUseJakartaEE10Websocket() {
-        return getBooleanProperty(FELIX_JAKARTA_EE10_WEBSOCKET_ENABLE, false);
+    public boolean isUseJakartaWebsocket() {
+        return getBooleanProperty(FELIX_JAKARTA_WEBSOCKET_ENABLE, false);
     }
 
     /**
      * Returns <code>true</code> if jetty websocket is configured to be used (
-     * {@link #FELIX_JETTY_EE10_WEBSOCKET_ENABLE})
+     * {@link #FELIX_JETTY_WEBSOCKET_ENABLE})
      */
-    public boolean isUseJettyEE10Websocket() {
-        return getBooleanProperty(FELIX_JETTY_EE10_WEBSOCKET_ENABLE, false);
+    public boolean isUseJettyWebsocket() {
+        return getBooleanProperty(FELIX_JETTY_WEBSOCKET_ENABLE, false);
     }
 
     public void reset()
